+++ src/plugins/Title/Title.tsx
@@ -4,11 +4,15 @@ import {Plugin, PluginWidgetProps} from '../../typings';
 import {cn} from '../../utils/cn';
 import {PLUGIN_ROOT_ATTR_NAME} from '../constants';
 
+import {RECCOMMENDED_LINE_HEIGHT_MULTIPLIER} from './constants';
+import type {PluginTitleSize, TitleFontParams} from './types';
+import {isCustomSize} from './utils';
+
 import './Title.scss';
 
 export interface PluginTitleProps extends PluginWidgetProps {
     data: {
-        size: 'l' | 'm' | 's' | 'xs';
+        size: PluginTitleSize | TitleFontParams;
         text: string;
         showInTOC: boolean;
     } & PluginWidgetProps['data'];
@@ -20,10 +24,25 @@ export class PluginTitle extends React.Component<PluginTitleProps> {
     render() {
         const {data} = this.props;
         const text = data.text ? data.text : '';
-        const size = data.size ? data.size : false;
+
+        const size = isCustomSize(data.size) ? false : data.size;
+        const styles =
+            isCustomSize(data.size) && data.size?.fontSize
+                ? {
+                      fontSize: data.size.fontSize,
+                      lineHeight: data.size.lineHeight ?? RECCOMMENDED_LINE_HEIGHT_MULTIPLIER,
+                  }
+                : undefined;
+
         const id = data.showInTOC && text ? encodeURIComponent(text) : undefined;
+
         return (
-            <div id={id} className={b({size})} {...{[PLUGIN_ROOT_ATTR_NAME]: 'title'}}>
+            <div
+                id={id}
+                style={styles}
+                className={b({size})}
+                {...{[PLUGIN_ROOT_ATTR_NAME]: 'title'}}
+            >
                 {text}
             </div>
         );