+++ src/plugins/Title/constants.ts
@@ -1,0 +1,26 @@
+import {PluginTitleSize, TitleFontParams} from './types';
+
+export const TITLE_DEFAULT_SIZES: Record<PluginTitleSize, TitleFontParams> = {
+    xl: {
+        fontSize: '32px',
+        lineHeight: '40px',
+    },
+    l: {
+        fontSize: '24px',
+        lineHeight: '28px',
+    },
+    m: {
+        fontSize: '20px',
+        lineHeight: '24px',
+    },
+    s: {
+        fontSize: '17px',
+        lineHeight: '24px',
+    },
+    xs: {
+        fontSize: '15px',
+        lineHeight: '20px',
+    },
+};
+
+export const RECCOMMENDED_LINE_HEIGHT_MULTIPLIER = 1.1;