+++ src/plugins/Title/index.ts
@@ -1,2 +1,4 @@
 export * from './Title';
 export {default as pluginTitle} from './Title';
+export {PluginTitleSize, TitleFontParams} from './types';
+export * from './constants';