+++ src/plugins/Title/utils.ts
@@ -1,0 +1,5 @@
+import type {PluginTitleSize, TitleFontParams} from './types';
+
+export function isCustomSize(size: PluginTitleSize | TitleFontParams): size is TitleFontParams {
+    return typeof size === 'object';
+}